Make dependency on `parking_lot` optional
Resolves #5 by marking the dependency on `parking_lot` as optional, with
the possibility to turn it off by disabling the default crate features.

The newtype pattern was central to this change, allowing both `std` and
`parking_lot` `RwLock` types to be used under the same internal
interface without any changes to existing code. The new `RwLock` wrapper
type has been defined as potentially (but not necessarily) poisonable
and as panicking when poisoned, which is adequate for the purposes of
this crate.

In addition, the crate docs have been updated to mention this new
optional feature.