Switch to u64 weights (#28)
This has the drawback of allowing for the possibility of weight overflows, but that's highly unlikely to be a problem in practice.

If it happens (and overflow checks disabled) it might cause the cache to get confused (e.g. allowing more items or infinite eviction loops) but it cannot lead to memory unsafety.